tag-expressions/python/tasks/py.requirements.txt
# ============================================================================
# INVOKE PYTHON PACKAGE REQUIREMENTS: For invoke tasks
# ============================================================================
# DESCRIPTION:
#    pip install -r <THIS_FILE>
#
# SEE ALSO:
#  * http://www.pip-installer.org/
# ============================================================================

invoke >= 0.21.0
six >= 1.11.0
path.py >= 10.1
pathlib;    python_version <= '3.4'
backports.shutil_which; python_version <= '3.3'

# -- CLEANUP SUPPORT: py.cleanup
pycmd
